Extract JWT generation into separate source file
