+++ src/codemagic/apple/app_store_connect/__init__.py
@@ -1,5 +1,5 @@
 from .api_client import AppStoreConnectApiClient
-from .api_client import IssuerId
-from .api_client import KeyIdentifier
 from .api_error import AppStoreConnectApiError
 from .api_session import AppStoreConnectApiSession
+from .type_declarations import IssuerId
+from .type_declarations import KeyIdentifier
