+++ src/codemagic/apple/app_store_connect/type_declarations.py
@@ -1,0 +1,17 @@
+from dataclasses import dataclass
+from typing import Dict
+from typing import List
+
+
+@dataclass
+class PaginateResult:
+    data: List[Dict]
+    included: List[Dict]
+
+
+class KeyIdentifier(str):
+    pass
+
+
+class IssuerId(str):
+    pass
